Oct 22 2021
Throughout this process, my code changed completely. I first started off by debugging all the mistakes that were in my code. I put the constructor in my kiosk when it was meant to go with the printer, I forgot to put (System. in) with my scanner, was using string shop multiple times before and didn't realize that I could take out the string once I use it again, and realized I had to add another shop.equals when comparing it to a lowercase y. When doing this quiz in class, I misunderstood and thought that the user could only have a maximum of 5 items, so I also ended up changing my while loop. In my previous code, I had the user type out which items they were buying and would repeat the question if they wanted to shop again. This would have been difficult to display in a receipt as a user might want more than one item of a product. This lead me to change my questions to ask them how much of each item they would like to buy. I had to then change this string input into a number in order to calculate the total number of items and the price of the items in total with the user's chosen quantity. In regards to the printer class, I am still struggling with it. My next steps would be to get the length of the 5 items and calculate the available spaces I have. I am having difficulty deciphering how to do this as I declared the items at the start as ints. I attempted to print out the receipt to the best of my ability, by adding \n and spaces.